refactor type tests into smaller test files
